+++ components/src/components/ModalHeader/ModalHeader.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { ModalHeader, ModalProvider } from "../lib";
-import ThemeProvider from "../components/contextProvider/ThemeProvider";
+import { ModalHeader, ModalProvider } from "../../lib";
+import ThemeProvider from "../contextProvider/ThemeProvider";
 
 const meta = {
   title: "ModalHeader",
